[PATCH] cciss: PCI BAR sizing fix
From: <mikem@beardog.cca.cpqcorp.net>

This patch eliminates the bad assumption that all of our PCI BARs will always
be 32-bits.  Tested against the 2.6.2 kernel.

This is required to support the Pinnacles architecture.  It is already in the
2.4 tree.

(This needs to be converted to use pci_request_regions())

+++ drivers/block/cciss.h
@@ -42,8 +42,8 @@ struct ctlr_info
 	char	firm_ver[4]; // Firmware version 
 	struct pci_dev *pdev;
 	__u32	board_id;
-	ulong   vaddr;
-	__u32	paddr;	
+	unsigned long vaddr;
+	unsigned long paddr;
 	unsigned long io_mem_addr;
 	unsigned long io_mem_length;
 	CfgTable_struct *cfgtable;
